refactor(workflow): use long IDs and self-contained SQL Server migration
- Widen workflow and related FK columns to bigint across entities, DTOs, services, and APIs.
- Add RefactorWorkflowTables migration with drop/recreate PKs and FKs so MigrateAsync works on fresh DBs.
- Update model snapshot and WorkflowApprovalStep configuration comment.
- Document optional DB verification queries under Database/Maintenance.

+++ Ettad.RequestManagement.Service/OrderItemTracking/OrderItemTrackingService.cs
@@ -82,7 +82,7 @@ namespace Ettad.RequestManagement.Service.OrderItemTracking
             }
         }
 
-        public async Task RecordFinalApprovalHistoryAsync(long orderId, int workflowApprovalStepId)
+        public async Task RecordFinalApprovalHistoryAsync(long orderId, long workflowApprovalStepId)
         {
             try
             {
