+++ Ettad.Workflow.Service/Dtos/WorkflowApprovalStepDto.cs
@@ -10,9 +10,9 @@ namespace Ettad.Workflows.Service.Dtos
 {
     public class WorkflowApprovalStepDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
-        public int WorkflowStepId { get; set; }         // link to WorkflowStep
+        public long WorkflowStepId { get; set; }         // link to WorkflowStep
         public long TargetRequestId { get; set; }
 
         public WorkflowType RequestType { get; set; }
@@ -30,7 +30,7 @@ namespace Ettad.Workflows.Service.Dtos
         public bool IsCurrent { get; set; }
 
         // When a step is returned for review, this tracks which step to return to after approval
-        public int? ReturnToStepId { get; set; }
+        public long? ReturnToStepId { get; set; }
 
         // Optional: For history tracking
         public RequestStatus? OldRequestStatus { get; set; }
@@ -43,7 +43,7 @@ namespace Ettad.Workflows.Service.Dtos
     public class CreateWorkflowApprovalStepDto
     {
         [Required]
-        public int WorkflowStepId { get; set; }
+        public long WorkflowStepId { get; set; }
 
         [Required]
         public long TargetRequestId { get; set; }
@@ -68,7 +68,7 @@ namespace Ettad.Workflows.Service.Dtos
     public class UpdateWorkflowApprovalStepDto
     {
         [Required]
-        public int WorkflowStepId { get; set; }
+        public long WorkflowStepId { get; set; }
 
         [Required]
         public long TargetRequestId { get; set; }
