+++ Ettad.Workflow.Service/Events/WorkflowStepApprovedEvent.cs
@@ -5,8 +5,8 @@ namespace Ettad.Workflows.Service.Events
 {
     public class WorkflowStepApprovedEvent : INotification
     {
-        public int WorkflowApprovalStepId { get; set; }
-        public int WorkflowStepId { get; set; }
+        public long WorkflowApprovalStepId { get; set; }
+        public long WorkflowStepId { get; set; }
         public long TargetRequestId { get; set; }
         public RequestType RequestType { get; set; }
         public string ApproverUserId { get; set; }
